+++ src/Plainion.GraphViz.Modules.Obfuscate/IObfuscator.cs
@@ -1,0 +1,6 @@
+﻿namespace Plainion.GraphViz.Modules.Obfuscate;
+
+internal interface IObfuscator
+{
+    string Obfuscate(string value);
+}